+++ omlete/src/main/java/omlete/mapper/ReviewMapper.java
@@ -11,7 +11,7 @@ public interface ReviewMapper {
 	int deleteReview(int reviewNo);
 
 	int updateReview(Review review);
-	int updateRate(int reviewStar, int reviewMember, int reviewKind,int reviewWork);
+	int updateRate(float reviewStar, int reviewMember, int reviewKind,int reviewWork);
 
 	Review selectReview(int reviewNo);
 	Review selectReviewByMemberno(int memberNo, int contentsNo);
